backend/requirements.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary

# Authentication & Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1

# API & Serialization
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==0.29.0
Flask-CORS==4.0.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing (opcional)
flask-testing==0.8.1
pytest==7.4.3
pytest-flask==1.3.0

#Validar Documentos do Brasil ( CPF )
validate-docbr==1.11.0